refactor(report): require via the shortest path
was using ../report/yadda instead of ./yadda :-/
