+++ src/report/anon/anonymizePath.js
