+++ src/report/anon/anonymizePathElement.js
