+++ test/report/anon/anonymize.spec.js
@@ -1,7 +1,7 @@
 const expect = require("chai").expect;
 const _clone = require("lodash/clone");
-const anonymize = require("../../../src/report/anonymous");
-const anonymizePathElement = require("../../../src/report/anonymous/anonymizePathElement");
+const anonymize = require("../../../src/report/anon");
+const anonymizePathElement = require("../../../src/report/anon/anonymizePathElement");
 const srcReport = require("./mocks/src-report.json");
 const srcReportWithWordlist = require("./mocks/src-report-wordlist.json");
 const fixtureReport = require("./fixtures/src-report.json");
