+++ test/report/anon/randomString.spec.js
@@ -1,7 +1,7 @@
 const expect = require("chai").expect;
-const randomString = require("../../../src/report/anonymous/randomString");
+const randomString = require("../../../src/report/anon/randomString");
 
-describe("report/anonymous/randomString", () => {
+describe("report/anon/randomString", () => {
   it("returns the empty string when passed the empty string", () => {
     expect(randomString("")).to.equal("");
   });
